+++ touchdown/tests/aws.py
@@ -280,6 +280,23 @@ class RecordedBotoCoreTest(unittest.TestCase):
         self.assertRaises(errors.NothingChanged, self.destroy_runner.execute)
 
 
+class StubberTestCase(unittest.TestCase):
+
+    TEST_CASE_GOAL = 'apply'
+
+    def setUp(self):
+        self.workspace = workspace.Workspace()
+        self.aws = self.workspace.add_aws(access_key_id='dummy', secret_access_key='dummy', region='eu-west-1')
+        self.goal = goals.create(
+            self.TEST_CASE_GOAL,
+            self.workspace,
+            ConsoleFrontend(interactive=False),
+            map=SerialMap
+        )
+        self.fixtures = ExitStack()
+        self.addCleanup(self.fixtures.close)
+
+
 class Stubber(BaseStubber):
     """Extends the stubber from botocore so that it always asserts that
     there are no leftover responses.
